feat: simplify the bare RN app so that it mirrors the Expo app
