--- ClassicImplementation/src/app/navigation/Screens.ts
@@ -1,14 +1,0 @@
-export enum Screens {
-  HOME = 'Home',
-  SCREEN_VIEWS = 'ScreenViews',
-  PRIVACY = 'Privacy',
-  WORKAROUNDS = 'Workarounds',
-  PAN_RESPONDER = 'PanResponder',
-  BASIC_SCREEN_VIEW = 'BasicScreenView',
-  PAGED_SCROLL_VIEW = 'PagedScrollView',
-  MODAL_SCREEN_VIEW = 'ModalScreenView',
-  PAGE_TAB_VIEW = 'PageTabView',
-  DYNAMIC_VARIABLES = 'DynamicVariables',
-  TRANSACTIONS = 'Transactions',
-  WEBVIEWS = 'Webviews',
-}
