--- ClassicImplementation/src/app/navigation/types.ts
@@ -1,27 +1,0 @@
-/**
- * Learn more about using TypeScript with React Navigation:
- * https://reactnavigation.org/docs/typescript/
- */
-
-import { NativeStackNavigationProp } from '@react-navigation/native-stack';
-import { Screens } from './Screens';
-
-export type RootStackParamList = {
-  [Screens.HOME]: undefined;
-  [Screens.SCREEN_VIEWS]: undefined;
-  [Screens.PRIVACY]: undefined;
-  [Screens.WORKAROUNDS]: undefined;
-  [Screens.PAN_RESPONDER]: undefined;
-  [Screens.BASIC_SCREEN_VIEW]: undefined;
-  [Screens.PAGED_SCROLL_VIEW]: undefined;
-  [Screens.MODAL_SCREEN_VIEW]: undefined;
-  [Screens.PAGE_TAB_VIEW]: undefined;
-  [Screens.DYNAMIC_VARIABLES]: undefined;
-  [Screens.TRANSACTIONS]: undefined;
-  [Screens.WEBVIEWS]: undefined;
-};
-
-export type RootNavigatorNavigationProps =
-  NativeStackNavigationProp<RootStackParamList>;
-
-export type ScreenNamesList = keyof RootStackParamList;
