--- ClassicImplementation/src/modules/Workarounds/views/Workarounds.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import { NavigationList } from '../../../shared/views/NavigationList/NavigationList';
-import { useWorkarounds } from './useWorkarounds';
-
-export const Workarounds: React.FunctionComponent = () => {
-  const { screensConfig } = useWorkarounds();
-
-  return <NavigationList screensConfig={screensConfig} />;
-};
